chore: add dumper interface & rename dumpers
