+++ common/src/java/com/zimbra/common/stats/Dumper.java
@@ -1,0 +1,7 @@
+package com.zimbra.common.stats;
+
+import java.util.concurrent.Callable;
+
+public interface Dumper extends Callable<Void> {
+  StatsDumperDataSource getDataSource();
+}
